requirements.txt
Python 3.6 
PyTorch 1.4.0  
Numpy 1.15.2 
gym 0.10.11 
opencv-Python
